+++ src/main/java/duke/exception/DukeCommandNotFoundException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class DukeCommandNotFoundException extends DukeException {
+    public DukeCommandNotFoundException () {
+        super("Command not found loll");
+    }
+}
