+++ src/main/java/duke/exception/DukeEmptyTaskException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class DukeEmptyTaskException extends DukeException {
+    public DukeEmptyTaskException () {
+        super("The task is empty loll");
+    }
+}
