+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,12 @@
+package duke.exception;
+
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+
+    @Override
+    public String toString() {
+        return "[OOPS!!]" + this.getMessage() + "\n";
+    }
+}
