+++ src/main/java/duke/ui/Ui.java
@@ -1,0 +1,14 @@
+package duke.ui;
+
+public class Ui {
+
+    public static void print(String s) {
+        System.out.println(s);
+    }
+
+    public void greet() {
+        print("Hello! I'm Duke\n" +
+                "     What can I do for you?");
+    }
+
+}
