+++ text-ui-test/EXPECTED.TXT
@@ -1,30 +1,30 @@
 Hello! I'm Duke
      What can I do for you?
-Got it. I've added this task:
-	[T][ ]  read book
+Got it. I've added this duke.task:
+	[T][ ] read book
 Now you have 1 tasks in the list.
-Nice! I've marked this task as done:
-	[T][X]  read book
-Got it. I've added this task:
+Nice! I've marked this duke.task as done:
+	[T][X] read book
+Got it. I've added this duke.task:
 	[D][ ] return book (by: June 6th)
 Now you have 2 tasks in the list.
-Got it. I've added this task:
-	[T][ ]  borrow book
+Got it. I've added this duke.task:
+	[T][ ] borrow book
 Now you have 3 tasks in the list.
-Got it. I've added this task:
+Got it. I've added this duke.task:
 	[D][ ] return book (by: Sunday)
 Now you have 4 tasks in the list.
-Got it. I've added this task:
+Got it. I've added this duke.task:
 	[E][ ] project meeting (from: Mon 2pm to 4pm)
 Now you have 5 tasks in the list.
-OK, I've marked this task as not done yet:
-	[T][ ]  read book
-Got it. I've added this task:
+OK, I've marked this duke.task as not done yet:
+	[T][ ] read book
+Got it. I've added this duke.task:
 	[D][ ] do homework (by: no idea :-p)
 Now you have 6 tasks in the list.
-1.[T][ ]  read book
+1.[T][ ] read book
 2.[D][ ] return book (by: June 6th)
-3.[T][ ]  borrow book
+3.[T][ ] borrow book
 4.[D][ ] return book (by: Sunday)
 5.[E][ ] project meeting (from: Mon 2pm to 4pm)
 6.[D][ ] do homework (by: no idea :-p)
